agents/aws_strands/04-agent-strands-session-store-sqlitedb-streamlit/requirements.txt
strands-agents>=1.2.0
strands-agents-tools>=0.2.3
fastapi
uvicorn